# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"BadRequestErrorResponse": {
				"description": "JSON response format for a \"400 Bad Request\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							400
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"fields": {
								"$ref": "#/components/schemas/FieldErrors"
							},
							"message": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 400,
					"error": {
						"message": "Error description.",
						"fields": [
							{
								"name": {
									"message": "Reason or description.",
									"value": "Any value or undefined."
								}
							}
						]
					}
				}
			},
...
			"NotFoundErrorResponse": {
				"description": "JSON response format for a \"404 Not Found\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							404
						],
						"nullable": false
					},
					"error": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 404,
					"error": {}
				}
			},
			"UUID": {
				"type": "string",
				"format": "uuid"
			},
...
			"Username": {
				"type": "string",
				"example": "user000",
				"pattern": "^[A-Za-z][\\w]{4,19}$"
			},
			"Password": {
				"type": "string",
				"example": "password",
				"pattern": "^[\\w]{8,19}$"
			},
...
			"USERS_NAME": {
				"type": "string",
				"example": "Alice"
			},
			"Role": {
				"enum": [
					"basic",
					"seller",
					"manager",
					"admin"
				],
				"type": "string"
			},
			"UserFullInfo": {
				"properties": {
					"userId": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"$ref": "#/components/schemas/USERS_NAME"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"userId",
					"name",
					"role",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUsersResult": {
				"description": "JSON response format for the \"GET /users\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"list": {
								"items": {
									"$ref": "#/components/schemas/UserFullInfo"
								},
								"type": "array"
							}
						},
						"required": [
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfile": {
				"properties": {
					"userId": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"$ref": "#/components/schemas/USERS_NAME"
					}
				},
				"required": [
					"userId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetSellersResult": {
				"description": "JSON response format for the \"GET /users\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"list": {
								"items": {
									"$ref": "#/components/schemas/UserProfile"
								},
								"type": "array"
							}
						},
						"required": [
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostUsersResult": {
				"description": "JSON response format for the \"POST /users\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							201
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserProfile"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserParams": {
				"description": "JSON request format to create a new user.",
				"properties": {
					"username": {
						"$ref": "#/components/schemas/Username"
					},
					"password": {
						"$ref": "#/components/schemas/Password"
					},
					"name": {
						"$ref": "#/components/schemas/USERS_NAME"
					}
				},
				"required": [
					"username",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUserProfileResult": {
				"description": "JSON response format for the \"GET /users/{userId}\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserProfile"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatchUserProfileResult": {
				"description": "JSON response format for the \"PATCH /users/{userId}\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserProfile"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserProfileParams": {
				"description": "JSON request format to update an existing user.",
				"properties": {
					"name": {
						"$ref": "#/components/schemas/USERS_NAME"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUserFullInfoResult": {
				"description": "JSON response format for the \"GET /users/{userId}/fullinfo\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserFullInfo"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatchUserFullInfoResult": {
				"description": "JSON response format for the \"PATCH /users/{userId}/fullinfo\" endpoint.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserFullInfo"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserFullInfoParams": {
				"description": "JSON request format to update an existing user's details.",
				"properties": {
					"name": {
						"$ref": "#/components/schemas/USERS_NAME"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "auth_token"
			}
		}
	},
	"info": {
		"title": "IndieLisboa Stock Management",
		"version": "1.0.0",
		"description": "Web API to view/edit product stock, register sales and manage user accounts.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Successfully returned a list of users.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersResult"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"error": {
												"fields": {
													"role": {
														"message": "should be one of the following; [...]",
														"value": "abc"
													}
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "If a search criteria is applied, only users with an **exact** match will be returned.",
				"summary": "Retrieve a list of users. You may specify search parameters.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Filter users by role.",
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Role"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Successfully created a new user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PostUsersResult"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestErrorResponse"
								}
							}
						}
					}
				},
				"summary": "Create a new user.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserParams"
							}
						}
					}
				}
			}
		},
		"/users/sellers": {
			"get": {
				"operationId": "GetSellers",
				"responses": {
					"200": {
						"description": "Successfully returned a list of users with the 'seller' role.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSellersResult"
								}
							}
						}
					}
				},
				"description": "Only retrieves users with the role 'seller'.",
				"summary": "Retrieve a list of users with the 'seller' role. You may specify search parameters.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"manager"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUserProfile",
				"responses": {
					"200": {
						"description": "Successfully returned the user's profile.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserProfileResult"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					}
				},
				"description": "Only the own user, or an account with sufficient privileges, may view this profile.",
				"summary": "Retrieve the user profile.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"basic"
						]
					}
				],
				"parameters": [
					{
						"description": "User's unique identifier.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchUserProfile",
				"responses": {
					"200": {
						"description": "Successfully updated the user's profile.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchUserProfileResult"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"error": {
												"fields": {
													"body.name": {
														"message": "invalid string value",
														"value": 0
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					}
				},
				"description": "Only the own user, or an account with sufficient privileges, may edit this profile.",
				"summary": "Update user profile.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"basic"
						]
					}
				],
				"parameters": [
					{
						"description": "User's unique identifier.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserProfileParams"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/fullinfo": {
			"get": {
				"operationId": "GetUserFullInfo",
				"responses": {
					"200": {
						"description": "Successfully returned the user's information.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserFullInfoResult"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					}
				},
				"summary": "Retrieve the user's information.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "User's unique identifier.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchUserFullInfo",
				"responses": {
					"200": {
						"description": "Successfully updated the user's information.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchUserFullInfoResult"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"error": {
												"fields": {
													"body.name": {
														"message": "invalid string value",
														"value": 0
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotFoundErrorResponse"
								}
							}
						}
					}
				},
				"summary": "Update user's information.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "User's unique identifier.",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserFullInfoParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}